First end-to-end working prototype of the TPS Client APIs
An end-to-end working demo of the TPS Client APIs which should work on any desktop
class machine, and should be easily portable into constrained embedded devices as
the Rust components do not require the Rust standard library (c_example requires a
C library currently)

There is a very basic demo of how to write a connector and TPS Service in Rust - ROT13.

- tps_error: Common Rust error handling for TPS crates
- tps_connector: Common Connector API definitions
- tps_client_common: Common definitions for TPS crates (no dependencies)
- tps_client_c: C language API (and example program) for the TPS Client API
- tps_client_api: The TPS Client API Rust library

At this point the code is of "demo" quality and intended to show the basic operation of
the APIs. Code should be thread-safe, but has not been extensively tested.

Signed-off-by: Jeremy O'Donoghue <quic_jodonogh@quicinc.com>