Fix #16: Remove vulnerable crates. No new features
The old version of clap crate used in the tps_cddl crate depended on
vulnerable or unmaintained crates. Updated to latest clap version,
which had a few breaking changes.

Note that cbindgen still depends on older versions of clap which
pulls in atty. As this is used only at build time the risk is deemed
acceptable.

Signed-off-by: Jeremy O'Donoghue <quic_jodonogh@quicinc.com>